+++ elfo/src/lib.rs
@@ -1,13 +1,17 @@
 #![warn(rust_2018_idioms, unreachable_pub)]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 pub use elfo_core::*;
 pub use elfo_macros::{message, msg};
 
 #[cfg(feature = "elfo-configurer")]
+#[cfg_attr(docsrs, doc(cfg(feature = "full")))]
 pub use elfo_configurer as configurer;
 #[cfg(feature = "elfo-logger")]
+#[cfg_attr(docsrs, doc(cfg(feature = "full")))]
 pub use elfo_logger as logger;
 #[cfg(feature = "test-util")]
+#[cfg_attr(docsrs, doc(cfg(feature = "test-util")))]
 pub use elfo_test as test;
 
 pub mod prelude {
@@ -17,5 +21,6 @@ pub mod prelude {
 #[deprecated(since = "0.1.1")]
 pub mod actors {
     #[cfg(feature = "elfo-configurer")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "full")))]
     pub use super::configurer;
 }