+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.1.16] - 2021-06-02
+### Added
+- docs: show features on docs.rs.
+
 ### Fixed
 - Terminate/restart actors more accurately.